Simplify API Program.cs with extension methods
