+++ FrontDesk/src/FrontDesk.Api/ServiceCollectionExtensions.cs
@@ -1,0 +1,39 @@
+﻿using System.Reflection;
+using FrontDesk.Core.Interfaces;
+using FrontDesk.Infrastructure.Messaging;
+using MassTransit;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace FrontDesk.Api
+{
+  public static class ServiceCollectionExtensions
+  {
+    public static void AddMessaging(this IServiceCollection services, ConfigurationManager configuration)
+    {
+      var messagingConfig = configuration.GetSection("RabbitMq");
+      services.Configure<RabbitMqConfiguration>(messagingConfig);
+      services.AddScoped<IMessagePublisher, MassTransitMessagePublisher>();
+
+      services.AddMassTransit(x =>
+      {
+        var rabbitMqConfiguration = messagingConfig.Get<RabbitMqConfiguration>();
+        x.SetKebabCaseEndpointNameFormatter();
+
+        x.AddConsumers(Assembly.GetExecutingAssembly());
+
+        x.UsingRabbitMq((context, cfg) =>
+        {
+          var port = (ushort)rabbitMqConfiguration.Port;
+          cfg.Host(rabbitMqConfiguration.Hostname, port, rabbitMqConfiguration.VirtualHost, h =>
+          {
+            h.Username(rabbitMqConfiguration.UserName);
+            h.Password(rabbitMqConfiguration.Password);
+          });
+
+          cfg.ConfigureEndpoints(context);
+        });
+      });
+    }
+  }
+}
